[22.08.17 여운] /post/detail controller

+++ src/main/java/com/ibini/my_books/hashtag/domain/HashtagDomain.java
@@ -7,7 +7,7 @@ import lombok.*;
 @AllArgsConstructor
 public class HashtagDomain {
     private int tagNo;
-    private int postNo;
+    private Long postNo;
     private String tagName;
     private String account;
 }
